linsolve: add CheckResidual operation
... and change how iteration is ended and how convergence is determined.

The motivation for this comes again from GMRES and from some convergence
failures observed during the rework of the testing framework.

Since GMRES does not form the residual as part of the iterative process,
there was previously a preliminary convergence test based on an estimate
of the residual norm. If this test succeeded, then the residual would be
formed explicitly and a final convergence test based on the residual
would be done. It could happen that the preliminary test would
succeed but the final test would fail and the method would restart with
this new residual. However, in some cases this situation could repeat
and the method would not make any progress, eventually exceeding the
maximum number of iterations.

This commit separates the convergence check from MajorIteration.
Convergence is to be checked only at CheckResidual and
CheckResidualNorm, and MajorIteration only indicates that the Method has
concluded its work for one iteration.