plotter/rings: new package migrated from biogo/graphics
This is a wholesale import of the code in github.com/biogo/graphics at
4122c1c03dca89559678b0e282180a3a5d757c19.

"gopkg.in/check.v1" is temporarily added to travis deps to confirm
build. This will be removed when std "testing" is used instead.

"github.com/biogo/biogo/feat" is temporarily added to travis deps
to allow the tests to pass and will be removed when the feature
basis type is altered.