linsolve: remove NormA from Settings
It was used in a convergence check after CheckResidual. The check relies
on Context.X to be up-to-date which is not a suitable requirements for
some methods like GMRES (which does not even provide the residual
directly). It is better to remove the field and the convergence check
that used it. If the user want a different convergence criteria, either
they have to code their own iterative loop or in the future we can
provide an iterface to provide custom criteria.