linsolve: remove CheckResidual operation
Currently there are two operations for checking convergence:
CheckResidual and CheckResidualNorm. The former exists for methods that
update the residual as part of the iterative process, while the latter
exists because for methods like GMRES forming the residual is expensive
and the residual norm is available cheaply. Since methods that do form
the residual during the iterations can easily compute the residual norm,
it is possible to remove the CheckResidual operation. This simplifies
the reverse communication interface because now all method must use
CheckResidualNorm and do not need to update the residual.