linsolve: remove Context.Residual
One of the previous commits removed the CheckResidual operation and so
methods are not required to update the Context.Residual field anymore
when they conclude an iteration. In effect this makes the field used
only for communicating the initial residual to the method. However,
Method.Init would be a much more natural place for doing this. And if we
do that, it makes all sense to pass also the initial X to Init along
with the residual. This also clarifies who is responsible for updating
the fields of Context: Methods update X, ResidualNorm, Src, and the
caller updates Converged and Dst.