linsolve: remove special handling of the first iteration
When we reuse slices in method, their contents is zeroed out as when a
slice is newly allocated. This makes possible to eliminate the 'first'
field from some methods and to simplify the code at the expense
of a few vector operations which are unnecessary in the first iteration
(because the operands are the zero vectors).